[DOCS/#14] 친구01 ~ 친구 02 UI 수정
