+++ app/src/main/java/com/example/teumteum/ui/friend/FollowingAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.teumteum.Friend
+package com.example.teumteum.ui.friend
 
 import android.view.LayoutInflater
 import android.view.View
