+++ app/src/main/java/com/example/teumteum/ui/friend/Friend02AcceptBottomSheetFragment.kt
@@ -1,4 +1,4 @@
-package com.example.teumteum.Friend
+package com.example.teumteum.ui.friend
 
 import android.graphics.Color
 import android.os.Bundle
