+++ app/src/main/java/com/example/teumteum/ui/friend/FriendRequestData.kt
@@ -1,4 +1,4 @@
-package com.example.teumteum.Friend
+package com.example.teumteum.ui.friend
 
 data class FriendRequestData(
     val name: String,
